create HelloPanel controller and moved everything out of the app controller
