+++ webapp/controller/App.controller.js
@@ -1,17 +1,8 @@
 sap.ui.define([
-    "sap/ui/core/mvc/Controller",
-    "sap/m/MessageToast"
-], (Controller, MessageToast) => {
+    "sap/ui/core/mvc/Controller"
+], (Controller) => {
     "use strict";
+    
     return Controller.extend("ui5.walkthrough.controller.App", {
-        onShowHello() {
-            // read msg from i18n model
-            const oBundle = this.getView().getModel("i18n").getResourceBundle();
-            const sRecipient = this.getView().getModel().getProperty("/recipient/name");
-            const sMsg = oBundle.getText("helloMsg", [sRecipient]);
-            
-            // show message
-            MessageToast.show(sMsg);
-        }
     });
 });
